Renamed reducer
