+++ src/contexts/RegisteredUsers.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useMemo, useReducer } from 'react';
 import registeredUsersReducer, {
   RegisteredUsersAction
-} from '../reducers/registeredUsers';
+} from '../reducers/registered-users';
 import { User } from '../types/API';
 
 export type RegisteredUsersContextType = {
